requirements.txt
Flask==3.1.0
torch==2.1.2
torchvision==0.16.2
numpy==1.26.4
prometheus-flask-exporter
grpcio-tools
grpcio
furl
pillow
supervisor